PIC-4372 split out blitzer test
